lambdas/image_processing/requirements.txt
Pillow==5.0.0
numpy==1.14.0
tensorflow==1.5.0

